Controller test and Created ReviewerController
